[!!!][TASK] Refactor display of heading and subheading (#4)
The display of heading and subheading has changed in the frontend. This means that the heading is now larger and the subheading is visually smaller.

It is essential to check whether the display of the headings is still correct after an upgrade. If the subheading is identical to the heading and is only used for design reasons, the subheading can be removed if necessary.